requirements.txt
asgiref==3.7.2
contourpy==1.1.1
cycler==0.12.1
Django==4.1
django-rest-api==0.1.5
djangorestframework==3.14.0
fonttools==4.43.1
kiwisolver==1.4.5
matplotlib==3.8.0
numpy==1.26.0
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
